pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyascii>=0.3.3",
    "datasets>=4.6.1",
    "dotenv>=0.9.9",
    "faker>=40.5.1",
    "google-cloud-storage>=3.9.0",
    "huggingface-hub>=1.5.0",
    "pillow>=12.1.1",
]
